django3/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-braces==1.14.0
django-crispy-forms==1.11.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
